requirements.txt
python-telegram-bot==12.8
gspread==3.6.0
oauth2client==4.1.2